backend/requirements/base.txt
# Base
pytz==2025.2
python-slugify==8.0.4
Pillow==11.3.0
psycopg2==2.9.10

# Django
Django==5.2.7

# Environment
django-environ==0.12.0

# Django REST Framework
djangorestframework==3.16.1
djangorestframework-recursive==0.1.2
django-filter==25.2

# JWT
PyJWT==2.10.1

# Passwords security
argon2-cffi==25.1.0

# Static files
whitenoise==5.2.0

# Cors Headers
django-cors-headers==4.9.0

# Swagger
drf-yasg==1.21.11

# Translations
django-modeltranslation==0.19.13

# Tools
django-extensions==4.1

# Utils
utm==0.8.1
beautifulsoup4==4.14.2
lxml==6.0.2

# Admin
django-admin-rangefilter==0.13.3

# Social Auth
social-auth-app-django==5.5.1
python-social-auth==0.3.6
social-auth-core==4.7.0
